requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
PyYAML==6.0.1
langdetect==1.0.9
transformers==4.35.2
torch==2.1.1
pydantic==2.5.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
